Infowindow migrator for builder #7973
