+++ app/controllers/carto/api/vizjson3_presenter.rb
@@ -1,4 +1,5 @@
 require_dependency 'carto/api/layer_vizjson_adapter'
+require_dependency 'carto/api/infowindow_migrator'
 require_dependency 'cartodb/redis_vizjson_cache'
 
 module Carto
@@ -425,6 +426,7 @@ module Carto
 
     class VizJSON3LayerPresenter
       include ApiTemplates
+      include InfowindowMigrator
 
       EMPTY_CSS = '#dummy{}'.freeze
 
@@ -462,8 +464,8 @@ module Carto
           {
             id:         @layer.id,
             type:       'CartoDB',
-            infowindow: whitelisted_attrs(with_template(@layer.infowindow, 'infowindows')),
-            tooltip:    whitelisted_attrs(with_template(@layer.tooltip, 'tooltips')),
+            infowindow: whitelisted_attrs(migrate_builder_infowindow(@layer.infowindow, mustache_dir: 'infowindows')),
+            tooltip:    whitelisted_attrs(migrate_builder_infowindow(@layer.tooltip, mustache_dir: 'tooltips')),
             legend:     @layer.legend,
             order:      @layer.order,
             visible:    @layer.public_values[:options]['visible'],
@@ -519,84 +521,6 @@ module Carto
                        end
 
         torque
-      end
-
-      MUSTACHE_ROOT_PATH = 'lib/assets/javascripts/cartodb3/mustache-templates'.freeze
-
-      def with_template(templated_element, mustache_dir)
-        return nil if templated_element.nil?
-
-        template = templated_element['template']
-        return templated_element if !template.nil? && !template.empty?
-
-        templated_sym = templated_element.deep_symbolize_keys
-
-        old_template_name = templated_sym[:template_name]
-
-        new_template_name = NEW_TEMPLATES_MAP.fetch(old_template_name, old_template_name)
-
-        if new_template_name == INFOWINDOW_COLOR_TEMPLATE
-          fixed_color = extract_color_from_old_template(templated_sym[:template_name])
-
-          template_content_path = "#{MUSTACHE_ROOT_PATH}/#{mustache_dir}/infowindow_color.jst.mustache"
-
-          templated_element[:template] = get_template(
-            new_template_name,
-            templated_sym[:template],
-            template_content_path).gsub('#35AAE5', fixed_color)
-
-          templated_element[:headerColor] = {
-            color: {
-              opacity: 1,
-              fixed: fixed_color
-            }
-          }
-
-          templated_element[:template_name] = new_template_name
-        else
-          templated_element[:template] = get_template(
-            templated_sym[:template_name],
-            templated_sym[:template],
-            "#{MUSTACHE_ROOT_PATH}/#{mustache_dir}/#{get_template_name(templated_sym[:template_name])}.jst.mustache")
-        end
-
-        templated_element
-      end
-
-      INFOWINDOW_COLOR_TEMPLATE = 'infowindow_color'.freeze
-
-      NEW_TEMPLATES_MAP = {
-        'infowindow_light' => 'infowindow_light',
-        'infowindow_dark' => 'infowindow_dark',
-        'infowindow_light_header_blue' => INFOWINDOW_COLOR_TEMPLATE,
-        'infowindow_light_header_yellow' => INFOWINDOW_COLOR_TEMPLATE,
-        'infowindow_light_header_orange' => INFOWINDOW_COLOR_TEMPLATE,
-        'infowindow_light_header_green' => INFOWINDOW_COLOR_TEMPLATE,
-        'infowindow_header_with_image' => 'infowindow_header_with_image'
-      }.freeze
-
-      COLOR_MAP = {
-        'blue' => '#35AAE5',
-        'green' => '#7FC97F',
-        'orange' => '#E68165',
-        'yellow' => '#E5C13D'
-      }.freeze
-
-      def extract_color_from_old_template(old_template_name)
-        COLOR_MAP.fetch(old_template_name.split('_').last, '#FABADA')
-      end
-
-      def get_template_name(name)
-        Carto::Layer::TEMPLATES_MAP.fetch(name, name)
-      end
-
-      def get_template(template_name, fallback_template, template_path)
-        if template_name.present?
-          path = Rails.root.join(template_path)
-          File.read(path)
-        else
-          fallback_template
-        end
       end
 
       def options_data
