+++ app/controllers/carto/builder/visualizations_controller.rb
@@ -21,7 +21,7 @@ module Carto
       def show
         @visualization_data = Carto::Api::VisualizationPresenter.new(@visualization, current_viewer, self).to_poro
         @layers_data = @visualization.layers.map do |l|
-          Carto::Api::LayerPresenter.new(l, with_style_properties: true).to_poro
+          Carto::Api::LayerPresenter.new(l, with_style_properties: true).to_poro(migrate_builder_infowindows: true)
         end
         @vizjson = generate_anonymous_map_vizjson3(@visualization, params)
         @analyses_data = @visualization.analyses.map { |a| Carto::Api::AnalysisPresenter.new(a).to_poro }
