+++ lib/assets/javascripts/cartodb3/editor/layers/layer-content-views/infowindow/infowindow-click-view.js
@@ -21,7 +21,7 @@ module.exports = InfowindowView.extend({
       }, {
         value: 'infowindow_color',
         infowindowIcon: require('./infowindow-icons/infowindow-color.tpl'),
-        label: _t('editor.layers.infowindow.style.infowindow_light_header_blue')
+        label: _t('editor.layers.infowindow.style.infowindow_color')
       }, {
         value: 'infowindow_header_with_image',
         infowindowIcon: require('./infowindow-icons/infowindow-image.tpl'),
